# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Agora CMS
  description: Central management system for Agora media playback devices
  version: 1.38.249
paths:
...
  /api/devices/{device_id}:
    get:
      summary: Get Device
      operationId: get_device_api_devices__device_id__get
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Device
      operationId: update_device_api_devices__device_id__patch
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Device
      operationId: delete_device_api_devices__device_id__delete
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/device-events:
    get:
      summary: List Device Events
      description: List device events visible to the current user, newest first.
      operationId: list_device_events_api_device_events_get
      parameters:
      - name: device_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
      - name: event_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Event Type
      - name: group_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Group Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          default: 100
          title: Limit
      - name: cursor_created_at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cursor Created At
      - name: cursor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Cursor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceEventOut'
                title: Response List Device Events Api Device Events Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/device-events/count:
    get:
      summary: Device Event Count
      description: Return total event count (for pagination / dashboards).
      operationId: device_event_count_api_device_events_count_get
      parameters:
      - name: device_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DeviceEventOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        device_name:
          type: string
          title: Device Name
        group_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Group Id
        group_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Group Ids
          default: []
        group_snapshots:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Group Snapshots
          default: []
        group_name:
          type: string
          title: Group Name
          default: ''
        event_type:
          type: string
          title: Event Type
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
        description:
          type: string
          title: Description
          default: ''
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - device_name
      - event_type
      - created_at
      title: DeviceEventOut
...
    DeviceOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        location:
          type: string
          title: Location
          default: ''
        status:
          $ref: '#/components/schemas/DeviceStatus'
        group_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Group Id
        group_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Name
        default_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Default Asset Id
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          title: Timezone
        firmware_version:
          type: string
          title: Firmware Version
        os_version:
          type: string
          title: Os Version
          default: ''
        device_type:
          type: string
          title: Device Type
          default: ''
        supported_codecs:
          type: string
          title: Supported Codecs
          default: ''
        storage_capacity_mb:
          type: integer
          title: Storage Capacity Mb
        storage_used_mb:
          type: integer
          title: Storage Used Mb
        last_seen:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Seen
        registered_at:
          type: string
          format: date-time
          title: Registered At
        is_online:
          type: boolean
          title: Is Online
          default: false
        is_upgrading:
          type: boolean
          title: Is Upgrading
          default: false
        upgrade_stuck:
          type: boolean
          title: Upgrade Stuck
          default: false
        ota_phase:
          anyOf:
          - type: string
          - type: 'null'
          title: Ota Phase
        ota_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Ota Label
        ota_pct:
          anyOf:
          - type: number
          - type: 'null'
          title: Ota Pct
        ota_bytes_done:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ota Bytes Done
        ota_bytes_total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ota Bytes Total
        playback_mode:
          anyOf:
          - type: string
          - type: 'null'
          title: Playback Mode
        playback_asset:
          anyOf:
          - type: string
          - type: 'null'
          title: Playback Asset
        pipeline_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Pipeline State
        display_connected:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Display Connected
        display_ports:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Display Ports
        has_active_schedule:
          type: boolean
          title: Has Active Schedule
          default: false
        cpu_temp_c:
          anyOf:
          - type: number
          - type: 'null'
          title: Cpu Temp C
        ip_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip Address
        ssh_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Ssh Enabled
        local_api_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Local Api Enabled
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        update_available:
          type: boolean
          title: Update Available
          default: false
        update_channel:
          type: string
          title: Update Channel
          default: stable
        uptime_seconds:
          type: integer
          title: Uptime Seconds
          default: 0
      type: object
      required:
      - id
      - name
      - status
      - firmware_version
      - storage_capacity_mb
      - storage_used_mb
      - registered_at
      title: DeviceOut
    DeviceStatus:
      type: string
      enum:
      - pending
      - adopted
      - orphaned
      title: DeviceStatus
    DeviceUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        status:
          anyOf:
          - $ref: '#/components/schemas/DeviceStatus'
          - type: 'null'
        group_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Group Id
        default_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Default Asset Id
        profile_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Profile Id
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          title: Timezone
        update_channel:
          anyOf:
          - type: string
          - type: 'null'
          title: Update Channel
      type: object
      title: DeviceUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Browser session cookie issued by POST /login.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Static API key issued via /api/keys; used for device-originated and automation callers.
servers:
- url: http://localhost:8080
  description: Local Docker Compose
security:
- sessionCookie: []
...